Making a silx.gui.plot._utils package
